requirements.txt
asgiref==3.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.4
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
requests==2.32.4
sqlparse==0.5.3
stripe==8.0.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
